+++ app/controllers/components_controller.rb
@@ -1,2 +1,3 @@
 class ComponentsController < ApplicationController
+  include Listable
 end