+++ app/controllers/elements_controller.rb
@@ -1,2 +1,3 @@
 class ElementsController < ApplicationController
+  include Listable
 end