+++ app/controllers/results_controller.rb
@@ -1,2 +1,3 @@
 class ResultsController < ApplicationController
+  include Listable
 end