+++ app/controllers/statics_controller.rb
@@ -1,0 +1,9 @@
+class StaticsController < ApplicationController
+  def index
+    render locals: {
+      category: params[:category].presence || "MEN",
+      season: params[:season].presence || "2016-17",
+    }
+      
+  end
+end