+++ app/decorators/result_decorator.rb
@@ -37,5 +37,6 @@ class ResultDecorator < EntryDecorator
   end
 
   decorate_as_score(:free_pcs, :free_deductions, :free_bv)
+  decorate_as_score(:total_bv)
 end