+++ db/migrate/002_create_competitions.rb
@@ -27,7 +27,8 @@ class CreateCompetitions < ActiveRecord::Migration[5.1]
       t.float :points
       t.integer :short_ranking
       t.integer :free_ranking
-
+      t.float :total_bv
+      
       t.belongs_to :competition
       t.references :skater
     end